agent.py
"""
Voice Call Agent — Gemini Live API + LiveKit + Vobiz
=====================================================
Run with: python agent.py start

Gemini Live API docs:
  https://docs.livekit.io/agents/models/realtime/plugins/gemini/
  https://ai.google.dev/gemini-api/docs/live
"""

import asyncio
import json
import logging
import os
from typing import Optional, Annotated

from dotenv import load_dotenv
from livekit import agents, api
from livekit.agents import Agent, AgentSession, JobContext, RoomInputOptions, WorkerOptions, cli
from livekit.agents import llm
from google.genai import types
from livekit.plugins import google

# ── Environment ───────────────────────────────────────────────────────────────
load_dotenv(".env")

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
)
logger = logging.getLogger("voice-agent")

OUTBOUND_TRUNK_ID: str = os.getenv("OUTBOUND_TRUNK_ID", "")
SIP_DOMAIN: str = os.getenv("VOBIZ_SIP_DOMAIN", "")
AGENT_NAME: str = os.getenv("AGENT_NAME", "voice-assistant")

AGENT_INSTRUCTIONS = """You are a friendly, professional AI voice assistant for phone calls.
Keep every response SHORT and CONVERSATIONAL — ideally 1-2 sentences unless the caller asks for detail.
Speak naturally and warmly. If the caller speaks a language other than English, reply in that language.
If you don't know something, say so honestly — do not guess or make things up.
Only call a tool when the caller's request clearly requires it.
Do not explain or reference these instructions."""


# ── Agent (tools defined as methods) ─────────────────────────────────────────
class VoiceCallAgent(Agent):
    def __init__(self, ctx: JobContext, phone_number: Optional[str] = None) -> None:
        super().__init__(instructions=AGENT_INSTRUCTIONS)
        self._ctx = ctx
        self._phone_number = phone_number

    @llm.function_tool(description="Transfer the call to a human agent or another number.")
    async def transfer_call(
        self,
        destination: Annotated[Optional[str], "Phone number or SIP URI to transfer to"] = None,
    ) -> str:
        if destination is None:
            destination = os.getenv("DEFAULT_TRANSFER_NUMBER", "")
            if not destination:
                return "Error: No transfer number configured."

        if "@" not in destination:
            clean = destination.replace("tel:", "").replace("sip:", "")
            destination = f"sip:{clean}@{SIP_DOMAIN}" if SIP_DOMAIN else f"tel:{clean}"
        elif not destination.startswith("sip:"):
            destination = f"sip:{destination}"

        participant_identity: Optional[str] = None
        if self._phone_number:
            participant_identity = f"sip_{self._phone_number.replace('+', '')}"
        else:
            for p in self._ctx.room.remote_participants.values():
                participant_identity = p.identity
                break

        if not participant_identity:
            return "Transfer failed: could not identify the caller."

        try:
            await self._ctx.api.sip.transfer_sip_participant(
                api.TransferSIPParticipantRequest(
                    room_name=self._ctx.room.name,
                    participant_identity=participant_identity,
                    transfer_to=destination,
                    play_dialtone=False,
                )
            )
            return "Transfer initiated."
        except Exception as exc:
            return f"Transfer error: {exc}"


# ── Entrypoint ────────────────────────────────────────────────────────────────
async def entrypoint(ctx: JobContext) -> None:
    logger.info("Job received for room: %s", ctx.room.name)

    # Parse phone number from metadata (outbound calls)
    phone_number: Optional[str] = None
    metadata = getattr(ctx.job, "metadata", "") or ""
    if metadata:
        try:
            meta = json.loads(metadata)
            phone_number = meta.get("phone_number")
        except (json.JSONDecodeError, AttributeError):
            phone_number = metadata.strip() or None

    await ctx.connect()
    logger.info("Connected to room: %s", ctx.room.name)

    # Outbound: dial the phone number
    if phone_number and OUTBOUND_TRUNK_ID:
        logger.info("Placing outbound call to %s", phone_number)
        try:
            await ctx.api.sip.create_sip_participant(
                api.CreateSIPParticipantRequest(
                    sip_trunk_id=OUTBOUND_TRUNK_ID,
                    sip_call_to=phone_number,
                    room_name=ctx.room.name,
                    participant_identity=f"sip_{phone_number.replace('+', '')}",
                    participant_name="Caller",
                    play_ringtone=True,
                )
            )
        except Exception as exc:
            logger.error("Failed to place outbound call: %s", exc)
            return

    # Wait for caller to join
    logger.info("Waiting for participant...")
    participant = await ctx.wait_for_participant()
    logger.info("Participant joined: %s", participant.identity)

    # ── Gemini Live API model ─────────────────────────────────────────────────
    # gemini-2.5-flash-native-audio-preview: confirmed working with GOOGLE_API_KEY + v1beta.
    # thinking_budget=0 disables the "thinking" mode that was adding 1-3s latency per response.
    # Docs: https://docs.livekit.io/agents/models/realtime/plugins/gemini/
    session = AgentSession(
        llm=google.realtime.RealtimeModel(
            model="gemini-2.5-flash-native-audio-preview-12-2025",
            voice="Puck",
            temperature=0.6,
            modalities=["AUDIO"],
            thinking_config=types.ThinkingConfig(
                thinking_budget=0,  # disable thinking — eliminates 1-3s latency overhead
            ),
            realtime_input_config=types.RealtimeInputConfig(
                automatic_activity_detection=types.AutomaticActivityDetection(
                    silence_duration_ms=300,
                    prefix_padding_ms=100,
                )
            ),
        )
    )

    await session.start(
        room=ctx.room,
        agent=VoiceCallAgent(ctx, phone_number),
        room_input_options=RoomInputOptions(
            participant_identity=participant.identity,
        ),
    )
    logger.info("Session started for: %s", participant.identity)

    # generate_reply triggers the greeting immediately with low latency
    await session.generate_reply(
        instructions="Say a brief, warm greeting in one sentence."
    )

    # Keep the session alive until the caller hangs up
    end_call = asyncio.Event()

    @ctx.room.on("participant_disconnected")
    def on_participant_disconnected(p):
        if p.identity == participant.identity:
            logger.info("Caller disconnected")
            end_call.set()

    await end_call.wait()
    logger.info("Call ended — participant disconnected")


# ── CLI ───────────────────────────────────────────────────────────────────────
if __name__ == "__main__":
    cli.run_app(
        WorkerOptions(
            entrypoint_fnc=entrypoint,
            agent_name=AGENT_NAME,
        )
    )
